refactor: fix parameter priority and export kernel interface functions for module VJP support
  This commit addresses a critical bug where module operations were ignoring
  config/autotuned values in favor of hardcoded parameter defaults, and adds
  proper exports of kernel interface functions to support platform-specific
  custom gradient implementations in module wrappers.

  ## Core Changes

  ### Parameter Priority Fix
  - Remove block_size parameter from native_sparse_attention module - now ONLY reads from config
  - Remove chunk_size_q/chunk_size_k parameters from flash_attention module - now ONLY reads from config
  - Fix blocksparse_attention to use cfg.q_blocksize/kv_blocksize instead of user parameters
  - Ensure autotuned/heuristic config values are respected instead of being overridden by defaults

  ### Kernel Interface Exports
  Export forward/backward interface functions from kernel __init__.py for module use:
  - Triton: flash_attention_gpu_fwd/bwd, native_sparse_attention_gpu_fwd/bwd,
    recurrent_gpu_fwd/bwd, blocksparse_attention_bhtd_fwd/bwd, mean_pooling_gpu_fwd/bwd
  - Pallas/TPU: flash_attention_tpu_fwd/bwd, ring_attention_tpu_fwd/bwd
  - XLA: native_sparse_attention_xla_fwd/bwd, recurrent_xla_fwd/bwd, ring_attention_xla_fwd/bwd

  ### Module Operation Enhancements
  - Add platform-specific fwd_with_residuals_* and vjp_* methods to support custom gradients
    across different backends (GPU/Triton, TPU/Pallas, XLA)
  - Add platform-specific candidate_cfgs_* methods (gpu/tpu/xla) for better autotuning
  - Add return_state parameter to recurrent-style operations (recurrent, gla, lightning_attention)
  - Optimize XLA operations to return empty candidate_cfgs() lists (no tunable parameters)

  ### Configuration System
  - Replace generic KernelConfig with typed config classes exported from operations:
    AttentionConfig, BlockSparseAttentionConfig, FlashAttentionConfig,
    NativeSparseAttentionConfig, RecurrentAttentionConfig, etc.
  - Export config classes from ejkernel.modules.operations
  - Deprecate create_default_executor() with migration guidance

  ### Bug Fixes
  - Fix dtype handling in ragged_decode_attention XLA implementation
  - Fix executor _cfg parameter extraction and passing

  ## Breaking Changes
  - Block size parameters removed from flash_attention and native_sparse_attention user APIs
  - create_default_executor() is now deprecated (use Executor(ConfigSelectorChain(...)) directly)

+++ ejkernel/kernels/_xla/ragged_decode_attention/_xla_impl_fwd.py
@@ -195,6 +195,10 @@ def flash_attention_block(
 
     o_new = (correction_prev * l_prev * o_prev + l_curr * o_curr) / l_new_safe
 
+    o_new = o_new.astype(o_prev.dtype)
+    m_new = m_new.astype(m_prev.dtype)
+    l_new = l_new.astype(l_prev.dtype)
+
     return (o_new, m_new, l_new), None
 
 
@@ -248,9 +252,9 @@ def ragged_flash_attention_xla(
 
     num_blocks = (kv_len + block_size - 1) // block_size
 
-    output_init = jnp.zeros_like(query)
-    max_logits_init = jnp.full((batch_size, q_len, num_heads, 1), -jnp.inf)
-    normalizer_init = jnp.zeros((batch_size, q_len, num_heads, 1))
+    output_init = jnp.zeros_like(query, dtype=query.dtype)
+    max_logits_init = jnp.full((batch_size, q_len, num_heads, 1), -jnp.inf, dtype=query.dtype)
+    normalizer_init = jnp.zeros((batch_size, q_len, num_heads, 1), dtype=query.dtype)
 
     pad_len = num_blocks * block_size - kv_len
     if pad_len > 0:
